make wrapper compatible with Python3
These changes make it possible to compile the wrapper under Python3 again.
All numpy references had to be removed due to incompatible API changes.
The function MatterProbPDGArr was removed entirely, as it is superfluous
at this point. The function SetParametersArr now takes a regular
Python list instead of a numpy array as input.